+++ core/src/main/java/org/seedstack/seed/core/internal/configuration/tool/DetailPrinter.java
@@ -5,7 +5,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package org.seedstack.seed.core.internal.tools.config;
+package org.seedstack.seed.core.internal.configuration.tool;
 
 import com.google.common.base.Strings;
 import org.fusesource.jansi.Ansi;