+++ core/src/main/java/org/seedstack/seed/core/internal/diagnostic/tool/ErrorsTool.java
@@ -5,7 +5,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package org.seedstack.seed.core.internal.tools.errors;
+package org.seedstack.seed.core.internal.diagnostic.tool;
 
 import io.nuun.kernel.api.plugin.InitState;
 import io.nuun.kernel.api.plugin.context.InitContext;